+++ srcs/ray/hit/hit_sphere.c
@@ -5,28 +5,12 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/15 22:03:31 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/22 14:29:10 by mfukui           ###   ########.fr       */
+/*   Created: 2025/04/22 16:25:33 by mfukui            #+#    #+#             */
+/*   Updated: 2025/04/22 16:25:36 by mfukui           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
-
-static float	solve_quadratic(float a, float b, float discriminant)
-{
-	float	t1;
-	float	t2;
-
-	if (discriminant < 0)
-		return (-1.0f);
-	t1 = (-b - sqrtf(discriminant)) / (2.0f * a);
-	t2 = (-b + sqrtf(discriminant)) / (2.0f * a);
-	if (t1 > 0.001f)
-		return (t1);
-	else if (t2 > 0.001f)
-		return (t2);
-	return (-1.0f);
-}
 
 static bool	set_sphere_quadratic(t_quadratic *q,
 		t_vector dir, t_vector oc, t_sphere *sphere)
