+++ srcs/utils/printers/rt_print_cylinder.c
@@ -6,7 +6,7 @@
 /*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/19 17:49:34 by tookuyam          #+#    #+#             */
-/*   Updated: 2025/04/22 01:09:54 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 16:43:57 by mfukui           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,5 +24,6 @@ void	rt_print_cylinder(const t_cylinder *cylinder)
 	printf("radius: %f\n", cylinder->radius);
 	rt_print_indent(PRINT_INDENT);
 	printf("height: %f\n", cylinder->height);
-	printf("color: %d,%d,%d\n", cylinder->color.r, cylinder->color.g, cylinder->color.b);
+	printf("color: %d,%d,%d\n", cylinder->color.r,
+		cylinder->color.g, cylinder->color.b);
 }
