+++ srcs/vector/vec_cross.c
@@ -6,15 +6,15 @@
 /*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 22:03:31 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/21 19:58:01 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 16:44:40 by mfukui           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-t_vector vec_cross(t_vector a, t_vector b)
+t_vector	vec_cross(t_vector a, t_vector b)
 {
-	t_vector result;
+	t_vector	result;
 
 	result.x = a.y * b.z - a.z * b.y;
 	result.y = a.z * b.x - a.x * b.z;
