+++ srcs/vector/vec_mul.c
@@ -6,15 +6,15 @@
 /*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 22:03:31 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/20 13:23:41 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 16:44:48 by mfukui           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-t_vector vec_mul(t_vector a, t_vector b)
+t_vector	vec_mul(t_vector a, t_vector b)
 {
-	t_vector result;
+	t_vector	result;
 
 	result.x = a.x * b.x;
 	result.y = a.y * b.y;
